# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: InMobi Test Java - Guess Number Game API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      security: []
      summary: Refresh JWT access token using a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/forgot-password/request:
    post:
      tags: [Auth]
      operationId: requestForgotPasswordOtp
      security: []
      summary: Request a password reset OTP sent to email (simulated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: OTP generated and email simulated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordRequestApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/forgot-password/reset:
    post:
      tags: [Auth]
      operationId: resetPasswordWithOtp
      security: []
      summary: Reset password using OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
  schemas:
    BaseResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: Success
    ErrorApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required: [data]
          properties:
            data:
              nullable: true
              allOf:
                - $ref: '#/components/schemas/ApiErrorDetail'
...
    ForgotPasswordRequestApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/ForgotPasswordResponse'
    ResetPasswordApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/ForgotPasswordResponse'
...
    ApiErrorDetail:
      type: object
      properties:
        violations:
          type: array
          items:
            type: string
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          minLength: 1
          maxLength: 255
    ResetPasswordRequest:
      type: object
      required: [email, otp, newPassword]
      properties:
        email:
          type: string
          format: email
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          minLength: 1
          maxLength: 255
        otp:
          type: string
          minLength: 1
          maxLength: 10
        newPassword:
          type: string
          minLength: 6
          maxLength: 72
    ForgotPasswordResponse:
      type: object
      required: [email, message]
      properties:
        email:
          type: string
        message:
          type: string
...
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, tokenType]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
...
    TokenRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenRefreshApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/AuthResponse'